tsk tsk on me, go ahead and quell the two size_t comparison warnings (besides, I might get lucky and cause another conflict for dtremenak :) muahaha
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@15005 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/ServerPing.cxx
@@ -68,7 +68,7 @@ int ServerPing::calcLag()
 
 bool ServerPing::done()
 {
-  return (recieved == samples || (activepings.size() == samples && (TimeKeeper::getCurrent() - activepings.back().senttime) > timeout));
+  return (recieved == samples || ((int)activepings.size() == samples && (TimeKeeper::getCurrent() - activepings.back().senttime) > timeout));
 }
 
 void ServerPing::setAddress(const Address& addr, int port)
@@ -90,7 +90,7 @@ void ServerPing::setInterval(double _interval)
 
 void ServerPing::doPings()
 { 
-  if ( activepings.size() < samples && (activepings.empty() || TimeKeeper::getCurrent() - activepings.back().senttime > interval) ) {
+  if ( (int)activepings.size() < samples && (activepings.empty() || TimeKeeper::getCurrent() - activepings.back().senttime > interval) ) {
     pingdesc pd;
     pd.senttime = TimeKeeper::getCurrent();
     sendPing(activepings.size());
